app launch behavior streamlined
the app now restores all documents from the previous session, but only if the general system preference option "Close windows when quitting an app" is disabled. If it is enabled, the app will show the welcome window (only if not disabled in the app's preferences), which it will also do when there are no documents from a previous session to be opened. When the system option is enabled and the welcome window is disabled, the app will show nothing after launch.